pyproject.toml
[tool.poetry]
name = "pdf2ppt"
version = "0.1.0"
description = ""
authors = ["Kang <caokang0720@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.2"
pdf2image = "^1.17.0"
python-pptx = "^0.6.23"
lxml = "^5.1.0"
pillow = "^10.2.0"
xlsxwriter = "^3.2.0"
pyqt5 = "^5.15.10"


[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
